Fix bugs#

+++ train.py
@@ -35,7 +35,7 @@ config = {
     "init_kl_coef": float(os.environ.get('INIT_KL_COEF', 0.05)),
     "target": int(os.environ.get('TARGET', 6)),
     "horizon": int(os.environ.get('HORIZON', 10000)),
-    "gamma": int(os.environ.get('GAMMA', 1)),
+    "gamma": float(os.environ.get('GAMMA', 1.0)),
     "lam": float(os.environ.get('LAM', 0.95)),
     "cliprange": float(os.environ.get('CLIPRANGE', 0.2)),
     "cliprange_value": float(os.environ.get('CLIPRANGE_VALUE', 0.2)),
